Added _BOpntsLoc()
This function located breakout points on the root locus and finds the gains associated with them. It uses the Characteristic equation N*D' - N'*D = 0 to solve for breakout points. Only breakout points on the real axis are kept. The gains are calculated with K = -D(s)/N(s), where s is the breakout point. Similarly, only the positive gains are kept.